navbar/mood/page

+++ app/src/main/java/com/example/jellyjamsapp/LeaderboardUser.kt
@@ -3,5 +3,5 @@ package com.example.jellyjamsapp
 data class LeaderboardUser(
     val userId: String,
     val username: String,
-    val score: Long
+    val score: Int
 )
